Change to single amqp connection
